Changed the statuses numbers to simplify status changes. Fixed bug where I was overwriting the status of an audit, with a downloaded, scheduled audit

+++ lib/pages/ListSchedulingPage/NewAuditDialog.dart
@@ -351,7 +351,7 @@ class _NewAuditDialogState extends State<NewAuditDialog> {
                     bool validated = validateEntry();
 
                     if (validated && timeInPastOK) {
-                      if (alreadyExisted) {
+                      if (alreadyExisted && !widget.followup) {
                         Provider.of<ListCalendarData>(context, listen: false)
                             .deleteCalendarItem(widget.calendarResult);
                       }
